requirements.txt
fastapi
uvicorn[standard]
pydantic
google-adk
google-genai
google-cloud-aiplatform
python-dotenv
slowapi
python-multipart
aiofiles
aiosqlite
asyncpg
psycopg2-binary